+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 _All notable changes to this project will be documented in this file._
 
+# 2020-07-13
+    - Redo the whole site with a custom, personal theme/template
+    - Added color-scheme-toggle plugin
+    - Added utterances for commenting
+
 # 2020-06-30
     - Migrated from custom jekyll build process to using github-pages gem